requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
async-timeout==4.0.3
attrs==23.1.0
better-profanity==0.7.0
black==23.9.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
discord.py==2.3.2
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
libsql-client==0.3.0
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.20
sqlalchemy-libsql==0.1.0
typing_extensions==4.7.1
urllib3==2.0.4
waifu-py==1.0.3
yarl==1.9.2
